-Categpries Table Enhancement Impl (#199)

+++ src/main/kotlin/com/realtor/plugins/data/model/category/Categories.kt
@@ -6,5 +6,5 @@ import kotlinx.serialization.Serializable
 data class Categories(
     val id: Long,
     val name: String,
-    val priority: Int,
+    val priority: Long,
 )
