+++ src/main/kotlin/com/realtor/plugins/data/table/category/CategoriesTable.kt
@@ -7,7 +7,7 @@ object CategoriesTable : Table(name = "Categories") {
 
     val id: Column<Long> = long("id").autoIncrement()
     val name: Column<String> = varchar("name", length = 51).uniqueIndex()
-    val priority: Column<Int> = integer("priority")
+    val priority: Column<Long> = long("priority")
 
     override val primaryKey: PrimaryKey = PrimaryKey(id)
 }
